+++ src/main/java/net/openhft/chronicle/wire/domestic/package-info.java
@@ -1,0 +1,16 @@
+/**
+ * This package and any and all sub-packages contains strictly internal classes reserved for Chronicle Software.
+ * Domestic classes shall <em>never</em> be used directly by end-user code. Instead, they are only available
+ * for use internally by other Chronicle libraries.
+ * <p>
+ * Specifically, the following actions (including, but not limited to) are not allowed
+ * on internal classes and packages:
+ *  <ul>
+ *      <li>Casting to</li>
+ *      <li>Reflection of any kind</li>
+ *      <li>Explicit Serialize/deserialize</li>
+ *  </ul>
+ * <p>
+ * The classes in this package and any sub-package are subject to change at any time for any reason.
+ */
+package net.openhft.chronicle.wire.domestic;
