+++ src/main/java/net/openhft/chronicle/wire/domestic/streaming/Streaming.adoc
