--- src/main/java/net/openhft/chronicle/wire/internal/streaming/Reduction.java
@@ -1,90 +1,0 @@
-package net.openhft.chronicle.wire.internal.streaming;
-
-import net.openhft.chronicle.core.annotation.NonNegative;
-import net.openhft.chronicle.wire.ExcerptListener;
-import net.openhft.chronicle.wire.MarshallableIn;
-import net.openhft.chronicle.wire.Wire;
-import net.openhft.chronicle.wire.internal.streaming.internal.ReductionUtil;
-import org.jetbrains.annotations.NotNull;
-
-import java.util.function.*;
-import java.util.stream.Collector;
-
-import static net.openhft.chronicle.core.util.ObjectUtils.requireNonNull;
-
-public interface Reduction<T> extends ExcerptListener {
-
-    /**
-     * Consumes an excerpt from the provided {@code wire} at the index at the provided {@code index}.
-     * <p>
-     * If this method throws an Exception, it is relayed to the call site.
-     * Therefore, care should be taken to minimise the probability of throwing Exceptions.
-     * <p>
-     * If this method is referenced as an {@link ExcerptListener} then the Reduction must be
-     * thread-safe.
-     **/
-    void onExcerpt(@NotNull Wire wire, @NonNegative long index);
-
-    /**
-     * Returns a view of the underlying reduction.
-     *
-     * @return Reduction view.
-     */
-    @NotNull
-    T reduction();
-
-    /**
-     * Accepts the input of the provided {@code tailer } and reduces (folds) the contents of it
-     * into this Reduction returning the last seen index or -1 if no index was seen.
-     * <p>
-     * This method can be used to initialise a Reduction before appending new values.
-     * <p>
-     * It is the responsibility of the caller to make sure no simultaneous appenders are using
-     * this Reduction during the entire fold operation.
-     *
-     * @param tailer to reduce (fold) from
-     * @return the last index seen or -1 if no index was seen
-     * @throws NullPointerException if the provided {@code tailer} is {@code null}
-     */
-    default long accept(@NotNull final MarshallableIn tailer) {
-        requireNonNull(tailer);
-        return ReductionUtil.accept(tailer, this);
-    }
-
-    // Basic static constructors
-
-    static <E> ReductionBuilder<E> of(@NotNull final DocumentExtractor<E> extractor) {
-        requireNonNull(extractor);
-        return new ReductionUtil.VanillaReductionBuilder<>(extractor);
-    }
-
-    static LongReductionBuilder ofLong(@NotNull final ToLongDocumentExtractor extractor) {
-        requireNonNull(extractor);
-        return new ReductionUtil.VanillaLongReductionBuilder(extractor);
-    }
-
-    static DoubleReductionBuilder ofDouble(@NotNull final ToDoubleDocumentExtractor extractor) {
-        requireNonNull(extractor);
-        return new ReductionUtil.VanillaDoubleReductionBuilder(extractor);
-    }
-
-    interface ReductionBuilder<E> {
-        <A, R> Reduction<R> collecting(@NotNull final Collector<E, A, ? extends R> collector);
-    }
-
-    interface LongReductionBuilder {
-
-        <A> Reduction<LongSupplier> reducing(@NotNull final Supplier<A> supplier,
-                                             @NotNull final ObjLongConsumer<A> accumulator,
-                                             @NotNull final ToLongFunction<A> finisher);
-
-    }
-
-    interface DoubleReductionBuilder {
-
-        <A> Reduction<DoubleSupplier> reducing(@NotNull final Supplier<A> supplier,
-                                               @NotNull final ObjDoubleConsumer<A> accumulator,
-                                               @NotNull final ToDoubleFunction<A> finisher);
-
-    }
-}
