--- src/main/java/net/openhft/chronicle/wire/internal/streaming/internal/IncubatorWarning.java
@@ -1,21 +1,0 @@
-package net.openhft.chronicle.wire.internal.streaming.internal;
-
-import net.openhft.chronicle.core.Jvm;
-
-public final class IncubatorWarning {
-
-    static {
-        Jvm.warn().on(IncubatorWarning.class,
-                "The incubating features are subject to change at any time for any reason and shall\n" +
-                        " not be used in production code. See net.openhft.chronicle.queue.incubator.package-info.java for details.");
-    }
-
-    // Suppresses default constructor, ensuring non-instantiability.
-    private IncubatorWarning() {
-    }
-
-    public static void warnOnce() {
-        // trigger static code
-    }
-
-}
